requirements.txt
boto3==1.9.197
Django==2.2.4
django-cors-headers==3.0.2
django-dotenv==1.4.2
django-filter==2.2.0
djangoql==0.13.0
django-import-export==1.2.0
django-rest-auth==0.9.5
django-storages==1.7.1
djangorestframework==3.10.1
dynamic-rest==1.9.4
graphene-django==2.5.0
pip==19.2.1
psycopg2-binary==2.8.3
PyYAML==5.1.1
setuptools==41.0.1